relocate stopAllPkgContainers
